Make JsonJobResultEntry an internal class of FilesystemJobResultStore
