+++ flink-runtime/src/main/java/org/apache/flink/runtime/highavailability/FileSystemJobResultStore.java
@@ -18,6 +18,7 @@
 
 package org.apache.flink.runtime.highavailability;
 
+import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.common.JobID;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.HighAvailabilityOptions;
@@ -25,9 +26,17 @@ import org.apache.flink.core.fs.FileStatus;
 import org.apache.flink.core.fs.FileSystem;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.runtime.jobmaster.JobResult;
+import org.apache.flink.runtime.rest.messages.json.JobResultDeserializer;
+import org.apache.flink.runtime.rest.messages.json.JobResultSerializer;
 import org.apache.flink.util.Preconditions;
 
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonCreator;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonIgnore;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonProperty;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -107,7 +116,7 @@ public class FileSystemJobResultStore extends AbstractThreadsafeJobResultStore {
     public void createDirtyResultInternal(JobResultEntry jobResultEntry) throws IOException {
         final Path path = constructDirtyPath(jobResultEntry.getJobId());
         OutputStream os = fileSystem.create(path, FileSystem.WriteMode.NO_OVERWRITE);
-        mapper.writeValue(os, new JobResultEntry.JsonJobResultEntry(jobResultEntry));
+        mapper.writeValue(os, new JsonJobResultEntry(jobResultEntry));
     }
 
     @Override
@@ -147,15 +156,54 @@ public class FileSystemJobResultStore extends AbstractThreadsafeJobResultStore {
                 if (!s.isDir()) {
                     String path = s.getPath().getPath();
                     if (path.contains(DIRTY_SUFFIX)) {
-                        JobResultEntry.JsonJobResultEntry jre =
+                        JsonJobResultEntry jre =
                                 mapper.readValue(
-                                        fileSystem.open(s.getPath()),
-                                        JobResultEntry.JsonJobResultEntry.class);
+                                        fileSystem.open(s.getPath()), JsonJobResultEntry.class);
                         dirtyResults.add(jre.getJobResult());
                     }
                 }
             }
         }
         return dirtyResults;
+    }
+
+    /**
+     * Wrapper class around {@link JobResultEntry} to allow for serialization of a schema version,
+     * so that future schema changes can be handled in a backwards compatible manner.
+     */
+    @JsonIgnoreProperties(
+            value = {JsonJobResultEntry.FIELD_NAME_VERSION},
+            allowGetters = true)
+    @VisibleForTesting
+    static class JsonJobResultEntry extends JobResultEntry {
+        private static final String FIELD_NAME_RESULT = "result";
+        private static final String FIELD_NAME_VERSION = "version";
+
+        JsonJobResultEntry(JobResultEntry entry) {
+            this(entry.getJobResult());
+        }
+
+        @JsonCreator
+        public JsonJobResultEntry(@JsonProperty(FIELD_NAME_RESULT) JobResult jobResult) {
+            super(jobResult);
+        }
+
+        @Override
+        @JsonProperty(FIELD_NAME_RESULT)
+        @JsonSerialize(using = JobResultSerializer.class)
+        @JsonDeserialize(using = JobResultDeserializer.class)
+        public JobResult getJobResult() {
+            return super.getJobResult();
+        }
+
+        @JsonIgnore
+        @Override
+        public JobID getJobId() {
+            return super.getJobId();
+        }
+
+        public int getVersion() {
+            return 1;
+        }
     }
 }
