+++ flink-runtime/src/main/java/org/apache/flink/runtime/highavailability/JobResultEntry.java
@@ -20,15 +20,7 @@ package org.apache.flink.runtime.highavailability;
 
 import org.apache.flink.api.common.JobID;
 import org.apache.flink.runtime.jobmaster.JobResult;
-import org.apache.flink.runtime.rest.messages.json.JobResultDeserializer;
-import org.apache.flink.runtime.rest.messages.json.JobResultSerializer;
 import org.apache.flink.util.Preconditions;
-
-import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonCreator;
-import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonProperty;
-import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 /**
  * {@code JobResultEntry} is the entity managed by the {@link JobResultStore}. It collects
@@ -48,39 +40,5 @@ public class JobResultEntry {
 
     public JobID getJobId() {
         return jobResult.getJobId();
-    }
-
-    @JsonIgnoreProperties(
-            value = {JsonJobResultEntry.FIELD_NAME_VERSION},
-            allowGetters = true)
-    /**
-     * Wrapper class around {@link JobResultEntry} to allow for serialization of a schema version,
-     * so that future schema changes can be handled in a backwards compatible manner.
-     */
-    public static class JsonJobResultEntry {
-        public static final String FIELD_NAME_RESULT = "result";
-        public static final String FIELD_NAME_VERSION = "version";
-
-        @JsonProperty(FIELD_NAME_RESULT)
-        @JsonSerialize(using = JobResultSerializer.class)
-        @JsonDeserialize(using = JobResultDeserializer.class)
-        private final JobResult jobResult;
-
-        public JsonJobResultEntry(JobResultEntry entry) {
-            this.jobResult = entry.getJobResult();
-        }
-
-        @JsonCreator
-        public JsonJobResultEntry(@JsonProperty(FIELD_NAME_RESULT) JobResult jobResult) {
-            this.jobResult = jobResult;
-        }
-
-        public JobResult getJobResult() {
-            return jobResult;
-        }
-
-        public int getVersion() {
-            return 1;
-        }
     }
 }
